Don't saturate lonePairs on carbon monoxide in solvation
Previously we would transform and lone pairs to biradicals in solvation thermo estimation.
Now that we have more specific atomTypes with lone pairs specified it is unnecessary,
although we may still lack the data for carbenes.

This commit avoids saturating lonePairs if the atomType of C is C2tc, which means the molecule is
carbon monoxide or isoelectronic to carbon monoxide.