Moved ensure_species_in_reaction to Reaction class
This method can be more widely applied to ensure species objects
are within the reaction and to convert the reactions that use
molecules. Since it only operates on the Reaction object, it makes
sense to keep it in the Reaction class.

To make the method more widely used, the method skips replacing
rxn.pairs if rxn.pairs is not instantiated.